Fix warnings

+++ src/phantasm-hardware-interface/d3d12/pools/resource_pool.hh
@@ -56,7 +56,7 @@ public:
             uint32_t width;
             uint32_t stride; // vertex/index size, structured buffer stride
 
-            bool is_access_in_bounds(uint32_t offset, uint32_t size) const { return offset + size <= width; }
+            bool is_access_in_bounds(size_t offset, size_t size) const { return offset + size <= size_t(width); }
         };
 
         struct image_info
@@ -97,7 +97,7 @@ public:
     [[nodiscard]] resource_node::image_info const& getImageInfo(handle::resource res) const { return internalGet(res).image; }
     [[nodiscard]] resource_node::buffer_info const& getBufferInfo(handle::resource res) const { return internalGet(res).buffer; }
 
-    bool isBufferAccessInBounds(handle::resource res, uint32_t offset, uint32_t size) const
+    bool isBufferAccessInBounds(handle::resource res, size_t offset, size_t size) const
     {
         auto const& internal = internalGet(res);
         if (internal.type != resource_node::resource_type::buffer)
